+++ src/Exception/MalformedPatternException.php
@@ -1,8 +1,6 @@
 <?php
 namespace TRegx\Exception;
 
-use TRegx\RegexException;
-
 interface MalformedPatternException extends RegexException
 {
 }